Add files via upload
After finishing the wiring yesterday, I found that the laser sensor’s signal was very unstable and easily affected by natural light. So I decided to try a new photo-gate sensor to see if it would work better. I also completed the 3D-printed housings for the sensor and the motor.